+++ lib/appfuel/storage/repository.rb
@@ -38,21 +38,13 @@ module Appfuel
     # @param entity_name [String] domain name of the entity we are mapping
     # @param db_class [String] name of the database class used in mapping
     # @return [DbEntityMapper]
-    def self.mapping(domain_name, options = {}, &block)
-      dsl = MappingDsl.new(domain_name, options)
+    def self.mapping(domain_name, to:, model:, **opts, &block)
+      dsl = MappingDsl.new(domain_name, to: to, model: model, **opts)
       dsl.instance_eval(&block)
 
-      dsl.entries.each do |entry|
-        root      = entry.container_name || Appfuel.default_app_name
-        container = Appfuel.app_container(root)
-        mappings  = container['repository_mappings']
-
-        domain_name = entry.domain_name
-        mappings[domain_name] = {} unless mappings.key?(domain_name)
-
-        entries = mappings[domain_name]
-        entries[entry.domain_attr] = entry
-      end
+      container = Appfuel.app_container(dsl.container_name)
+      mappings  = container['repository_mappings']
+      mappings.load(dsl)
     end
 
     def self.entity_builder(domain_name, type, opts = {}, &block)