+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file. (Pending ap
 ### Added
 - adding new repository type of aws dynamodb
 - adding initializers for aws dynamodb
+### Changed
+- storage mapping as simplified to a one-to-one between domain and storage model
 
 ## [[0.4.5]] (https://github.com/rsb/appfuel/releases/tag/0.4.5) 2017-06-29
 ### Fixed